feat(Refeicao): mudar peso para doble

+++ backend/CalculadoraCalorias.Application/DTOs/Responses/RefeicaoModeloResponse.cs
@@ -8,6 +8,6 @@ namespace CalculadoraCalorias.Application.DTOs.Responses
         public double? Proteinas { get; set; }
         public double? Carboidratos { get; set; }
         public double? Gorduras { get; set; }
-        public int PesoOriginal { get; set; }
+        public double PesoOriginal { get; set; }
     }
 }
